+++ scripts/test/Package.t
@@ -11,9 +11,10 @@ use strict;
 use warnings;
 use vars (qw($Self));
 
-use Kernel::System::Package;
 use File::Copy;
+use Kernel::System::Package;
 use Kernel::System::Cache;
+use Kernel::System::VariableCheck qw(:all);
 
 # create local objects
 my $PackageObject = Kernel::System::Package->new( %{$Self} );
@@ -1724,5 +1725,103 @@ if ( !$DeveloperSystem ) {
         '#14 TestPackageUninstallMerged PackageIsInstalled() - with flase',
     );
 }
+
+# PackageParse method basic test
+
+my $StringNoXML = 'Not a valid structure
+for a package file.';
+
+my $ResultStructure = 1;
+my %StructureFail = $PackageObject->PackageParse( String => $StringNoXML );
+$ResultStructure = 0 if !IsHashRefWithData( \%StructureFail );
+$Self->Is(
+    $ResultStructure,
+    0,
+    "#15 PackageParse() - Wrong package content",
+);
+
+my $StringInvalid = '
+<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01//EN"
+    "http://www.w3.org/TR/html4/strict.dtd">
+<html>
+  <head>
+    <title>Page title</title>
+  </head>
+  <body>
+    <div class="Main">
+        <p>This is a invalid content.</p>
+        <p>It can pass the XML parse.</p>
+        <p>But is not possible to retrieve an structure from it.</p>
+    </div>
+  </body>
+</html>
+';
+
+$ResultStructure = 1;
+%StructureFail   = $PackageObject->PackageParse( String => $StringInvalid );
+$ResultStructure = 0 if !IsHashRefWithData( \%StructureFail );
+$Self->Is(
+    $ResultStructure,
+    0,
+    "15 PackageParse() - Invalid package content.",
+);
+
+my $StringNormal = '<?xml version="1.0" encoding="utf-8" ?>
+    <otrs_package version="1.0">
+      <Name>TestPackage</Name>
+      <Version>1.0.1</Version>
+      <Vendor>OTRS AG</Vendor>
+      <URL>http://otrs.org/</URL>
+      <License>GNU GENERAL PUBLIC LICENSE Version 2, June 1991</License>
+      <ChangeLog>2013-08-14 New package (some test &lt; &gt; &amp;).</ChangeLog>
+      <Description Lang="en">A test package (some test &lt; &gt; &amp;).</Description>
+      <Description Lang="de">Ein Test Paket (some test &lt; &gt; &amp;).</Description>
+      <ModuleRequired Version="1.112">Encode</ModuleRequired>
+      <Framework>3.3.x</Framework>
+      <Framework>3.2.x</Framework>
+      <Framework>3.1.x</Framework>
+      <Framework>3.0.x</Framework>
+      <Framework>2.5.x</Framework>
+      <Framework>2.4.x</Framework>
+      <Framework>2.3.x</Framework>
+      <Framework>2.2.x</Framework>
+      <Framework>2.1.x</Framework>
+      <Framework>2.0.x</Framework>
+      <BuildDate>2005-11-10 21:17:16</BuildDate>
+      <BuildHost>yourhost.example.com</BuildHost>
+      <Filelist>
+        <File Location="Test" Permission="644" Encode="Base64">aGVsbG8K</File>
+        <File Location="var/Test" Permission="644" Encode="Base64">aGVsbG8K</File>
+        <File Location="bin/otrs.CheckDB.pl" Permission="755" Encode="Base64">aGVsbG8K</File>
+      </Filelist>
+    </otrs_package>
+';
+
+$ResultStructure = 0;
+my %StructureNormal = $PackageObject->PackageParse( String => $StringNormal );
+$ResultStructure = 1 if IsHashRefWithData( \%StructureNormal );
+$Self->Is(
+    $ResultStructure,
+    1,
+    "15 PackageParse() - Normal package content",
+);
+
+$Self->Is(
+    $StructureNormal{Name}->{Content},
+    'TestPackage',
+    "15 PackageParse() - Normal package content | Name ",
+);
+
+$Self->Is(
+    $StructureNormal{Version}->{Content},
+    '1.0.1',
+    "15 PackageParse() - Normal package content | Version ",
+);
+
+$Self->Is(
+    $StructureNormal{Vendor}->{Content},
+    'OTRS AG',
+    "15 PackageParse() - Normal package content | Vendor ",
+);
 
 1;